requirements.txt
asgiref==3.2.7
astroid==2.4.1
boto3==1.14.46
botocore==1.17.46
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-braces==1.14.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.10
whitenoise==5.0.1
wincertstore==0.2
wrapt==1.12.1
